Refactored middleware from router. Added logging.
